# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Construction Ledger API
  description: >
    REST API for the Construction Ledger app. Projects come in three kinds —
    'site' (construction sites, credit/debit), 'account' (Srishti Construction /
    Sri Srishti Construction, withdrawal-only, reflects credits tagged to them
    from site payments), and 'loan' (credit = disbursement, debit = repayment).
  version: 3.0.0
servers:
  - url: http://localhost:4000
    description: Local Docker dev
  # - url: https://your-backend-url.example.com
  #   description: Production
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in and receive a JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: admin }
                password: { type: string, format: password, example: changeme123 }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  token: { type: string, description: JWT, valid 30 days }
                  user:
                    type: object
                    properties:
                      user_id: { type: string, format: uuid }
                      username: { type: string }
                      name: { type: string }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, name]
              properties:
                username: { type: string, example: admin }
                password: { type: string, format: password, example: changeme123 }
                name: { type: string, example: Ishwar Singh }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  user_id: { type: string, format: uuid }
                  data:
                    type: object
                    properties:
                      username: { type: string }
                      name: { type: string }
        '400':
          description: Missing required fields or username already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/projects:
    get:
      tags: [Projects]
      summary: List all projects (sites + the 2 accounts + the loan project)
      description: >
        Each project now includes its own computed total_credit, total_debit,
        total_withdrawal, and balance (balance formula depends on kind — see
        the Project schema).
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }
    post:
      tags: [Projects]
      summary: Create a new site
      description: >
        Always creates kind='site'. NOTE — current implementation accepts an
        optional project_code in the body; if omitted, this will fail the
        database's NOT NULL constraint rather than auto-generating one.
        Confirm whether client-supplied codes are actually intended before
        relying on this.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_name]
              properties:
                project_name: { type: string, example: Ganga Niketan Site }
                project_code: { type: string, example: GANG-482910, description: "Optional. See description above — currently required in practice." }
                start_date: { type: string, format: date }
                status: { type: string, example: active }
      responses:
        '201':
          description: Site created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  project_id: { type: string, format: uuid }
                  data: { $ref: '#/components/schemas/Project' }
        '400':
          description: Missing project_name (or missing project_code — see note above)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/projects/company-summary:
    get:
      tags: [Projects]
      summary: Company-wide operational position
      description: >
        site credits − site debits − all account withdrawals − total loan
        repaid. Deliberately not the same as summing every project's own
        balance — this is "where do we actually stand after partners took
        their share and loan repayments are accounted for."
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Company summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      site_credits: { type: number, example: 17496641 }
                      site_debits: { type: number, example: 16047709 }
                      total_withdrawals: { type: number, example: 1697428 }
                      total_loan_repaid: { type: number, example: 72500 }
                      balance: { type: number, example: -321000 }
  /api/projects/{id}:
    get:
      tags: [Projects]
      summary: Get a single project by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: '#/components/schemas/Project' }
        '404':
          description: Project not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    patch:
      tags: [Projects]
      summary: Update a project's name, start date, or status
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                project_name: { type: string }
                start_date: { type: string, format: date }
                status: { type: string }
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: '#/components/schemas/Project' }
        '404':
          description: Project not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      tags: [Projects]
      summary: Delete a site (cascades to its transactions)
      description: Blocked with a 400 for kind='account' or kind='loan' — the 2 accounts and the loan project are structural and fixed.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '400':
          description: Attempted to delete a non-site project
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Project not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Obtain via POST /api/auth/login.
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Project:
      type: object
      properties:
        project_id: { type: string, format: uuid }
        project_code: { type: string, example: GANG-482910 }
        project_name: { type: string, example: Ganga Niketan Site }
        kind: { type: string, enum: [site, account, loan], example: site }
        start_date: { type: string, format: date }
        status: { type: string, example: active }
        created_at: { type: string, format: date-time }
        total_credit: { type: number, description: Now included directly on GET /api/projects list items }
        total_debit: { type: number }
        total_withdrawal: { type: number }
        balance: { type: number, description: "Formula depends on kind — see /api/projects/{id}/summary description" }
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string, example: "Descriptive error message" }
security:
  - bearerAuth: []